+++ FilePocket.Client/Features/Files/FileUrl.cs
@@ -7,6 +7,7 @@
         public static string GetFileInfo(Guid pocketId, Guid fileId) => $"api/pockets/{pocketId}/files/{fileId}/info";
         public static string GetImageThumbnail(Guid pocketId, Guid imageId, int size) => $"api/pockets/{pocketId}/files/{imageId}/thumbnail/{size}";
         public static string GetAll(Guid pocketId) => $"api/pockets/{pocketId}/files";
+        public static string GetAll(Guid pocketId, Guid folderId) => $"api/pockets/{pocketId}/folders/{folderId}/files";
         public static string DeleteFile(Guid pocketId, Guid fileId) => $"api/pockets/{pocketId}/files/{fileId}";
     }
 }
