+++ FilePocket.Client/Pages/Files/FolderFiles.razor
@@ -1,0 +1,3 @@
+﻿@page "/pockets/{PocketIdParam}/folders/{FolderIdParam}"
+
+<FilesAndFolders PocketId="_pocketId" FolderId="_folderId"></FilesAndFolders>
