+++ FilePocket.Client/Pages/Folders/CreateFolder.razor
@@ -1,5 +1,5 @@
-﻿@page "/pockets/{PocketIdParam}/folders/new"
-
+﻿@page "/pockets/{PocketIdParam}/folders/{FolderIdParam}/new"
+       
 @inject IJSRuntime JS
 @inject NavigationManager Navigation;
