+++ FilePocket.Client/Pages/Pockets/Pockets.razor
@@ -52,7 +52,7 @@
                             <tr>
                                 <th scope="row">@(i)</th>
                                 <td>
-                                    <a class="link-offset-2 link-underline link-underline-opacity-0 link-dark" href="/pockets/{@pocket.Id}/files">
+                                    <a class="link-offset-2 link-underline link-underline-opacity-0 link-dark" href="/pockets/@pocket.Id/files">
                                         @* <img src="./assets/img/pocket.png" width="24" class="me-1" /> *@
                                         <i class="bi bi-archive fs-5 me-1"></i>
                                         <div class="d-inline h5 mt-5">
