+++ FilePocket.Contracts/Repositories/IFileUploadSummaryRepository.cs
@@ -7,6 +7,8 @@ public interface IFileUploadSummaryRepository
 {
     Task<IEnumerable<FileUploadSummary>> GetAllByStorageIdAsync(Guid storageId, bool trackChanges = false);
 
+    Task<IEnumerable<FileUploadSummary>> GetAllByStorageIdAndFolderIdAsync(Guid storageId, Guid folderId, bool trackChanges = false);
+
     Task<IEnumerable<FileUploadSummary>> GetFilteredFilesAsync(FilesFilterOptionsModel filterOptionsModel, bool trackChanges = false);
 
     Task<int> GetFilteredCountAsync(FilesFilterOptionsModel filterOptionsModel, bool trackChanges = false);
